+++ news/static/news/js/event/ticket/email_copy.js
@@ -1,9 +1,21 @@
 let popupTimer;
 
-function delayPopup($popup) {
+function createPopup($elementToCreateFor, popupHtml) {
+    $elementToCreateFor
+        .popup({
+            html: popupHtml,
+            on: "manual",
+            exclusive: true,
+        })
+        .popup("show");
+    // Hide popup after 5 seconds
+    delayPopup($elementToCreateFor, 5000);
+}
+
+function delayPopup($popup, delayMillis) {
     popupTimer = setTimeout(function () {
-        $popup.popup('hide');
-    }, 4200);
+        $popup.popup("hide");
+    }, delayMillis);
 }
 
 $(".copy-token").click(function () {
@@ -11,19 +23,19 @@ $(".copy-token").click(function () {
     clearTimeout(popupTimer);
 
     const $input = $copyButton.closest(".emails-container").find(".copy-input");
-    // Select the text field
+    // Make the text field selected, to communicate to the user what text was copied
     $input.select();
+    const emails = $input.val();
 
     // Copy the text inside the text field
-    document.execCommand('copy');
-
-    $copyButton
-        .popup({
-            title: gettext("Successfully copied to clipboard!"),
-            on: 'manual',
-            exclusive: true,
-        })
-        .popup('show');
-    // Hide popup after 5 seconds
-    delayPopup($copyButton);
+    navigator.clipboard.writeText(emails).then(
+        // Clipboard successfully set
+        () => {
+            createPopup($copyButton, `<div class="ui green header">${gettext("Copied to clipboard!")}</div>`);
+        },
+        // Clipboard write failed
+        () => {
+            createPopup($copyButton, `<div class="ui red header">${gettext("Failed to copy to clipboard!")}</div>`);
+        },
+    );
 });