+++ util/templatetags/uri_tags.py
@@ -10,6 +10,9 @@ def get_absolute_uri_for_path(request, path: str):
     :param path: the path to append to the request's scheme and host
     :return: The absolute URI of the provided path.
     """
+    if not request:
+        return ""
+
     return request.build_absolute_uri(path)
 
 
@@ -19,4 +22,7 @@ def get_absolute_uri_no_query(request):
     :param request: the request object
     :return: The absolute URI of the requested page, without any query parameters.
     """
+    if not request:
+        return ""
+
     return get_absolute_uri_for_path(request, request.path)