+++ web/admin.py
@@ -1,7 +1,10 @@
 from django.contrib import admin
+from django_hosts import reverse_lazy
 
 
 class WebAdminSite(admin.AdminSite):
+    site_url = reverse_lazy('front_page')
+
     apps_listed_first = (
         'constance',